IGNITE-3303 Apache Flink Integration - Flink source to run a continuous query against one or multiple caches

+++ modules/flink/src/test/java/org/apache/ignite/source/flink/FlinkIgniteSourceSelfTest.java
@@ -147,7 +147,7 @@ public class FlinkIgniteSourceSelfTest extends GridCommonAbstractTest {
         final List<Integer> resultList = new ArrayList<>();
 
         while(myOutput.hasNext()){
-            Integer testKey = (Integer) myOutput.next().key();
+            Integer testKey = myOutput.next().key();
 
             X.println(">>> Printing iterator results: "+ testKey);
